+++ CHANGELOG.md
@@ -83,6 +83,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `--insert-before` now accepts a revset that resolves to an empty set when
   used with `--insert-after`. The behavior is similar to `--onto`.
 
+* Tags can now be pushed to a remote, using `jj git push`. To push all local tags,
+  use `jj git push --tags`. To push a specific tag, use `jj git push --tag NAME`.
+
 ### Fixed bugs
 
 * Broken symlink on Windows. [#6934](https://github.com/jj-vcs/jj/issues/6934).